Only override VERSION and REVMARK for manual GitHub Actions builds
For non-manual builds the "env" argument to the "Build Files" step will
set VERSION to "v" and REVMARK to the empty string which results in
the version of the PDF being incorrect. To avoid this problem use a
conditional step that updates $GITHUB_ENV for `workflow_dispatch` builds
instead of setting the variables unconditionally.

This PR also removes the Step <N> prefix to all the comments since it
does not add any explanatory value and makes it more difficult to update
the workflow file.

See https://github.com/riscv/riscv-cheri/issues/72

Signed-off-by: Alex Richardson <alexrichardson@google.com>